+++ Test/MatIF_GD.py
@@ -289,6 +289,11 @@ if __name__ == '__main__':
     for i in range(len(master_files)):
         master_files[i] = master_files[i].split("/")[-1].split(".")[0]
     
+    if os.path.exists("Results/MatIF_Uni.csv"):
+        MatIF_Uni = pd.read_csv("Results/MatIF_Uni.csv")
+        done_files = MatIF_Uni["Filename"].to_numpy()
+        master_files = [item for item in master_files if item not in done_files]
+    
     master_files.sort()
     parameters = []
     
@@ -303,14 +308,14 @@ if __name__ == '__main__':
     frr=open("GD_ReRun/MatIF.csv", "w")
     frr.write('Filename,ContaminationFraction,NumLearners,NumObservationsPerLearner\n')
     frr.close()
-    
-    frr=open("Results/MatIF_Uni.csv", "w")
-    frr.write('Filename,ContaminationFraction,NumLearners,NumObservationsPerLearner\n')
-    frr.close()
-    
-    frr=open("Results/MatIF_Bi.csv", "w")
-    frr.write('Filename,ContaminationFraction,NumLearners,NumObservationsPerLearner\n')
-    frr.close()
+    if os.path.exists("Results/MatIF_Uni.csv") == 0:
+        frr=open("Results/MatIF_Uni.csv", "w")
+        frr.write('Filename,ContaminationFraction,NumLearners,NumObservationsPerLearner\n')
+        frr.close()
+    if os.path.exists("Results/MatIF_Bi.csv") == 0:
+        frr=open("Results/MatIF_Bi.csv", "w")
+        frr.write('Filename,ContaminationFraction,NumLearners,NumObservationsPerLearner\n')
+        frr.close()
     
     for fname in master_files:
         isolationforest(fname, parameters, 0)
