+++ Test/MatOCSVM_GD.py
@@ -291,6 +291,11 @@ if __name__ == '__main__':
     for i in range(len(master_files)):
         master_files[i] = master_files[i].split("/")[-1].split(".")[0]
     
+    if os.path.exists("Results/MatOCSVM_Uni.csv"):
+        MatOCSVM_Uni = pd.read_csv("Results/MatOCSVM_Uni.csv")
+        done_files = MatOCSVM_Uni["Filename"].to_numpy()
+        master_files = [item for item in master_files if item not in done_files]
+    
     master_files.sort()
     parameters = []
 
@@ -317,13 +322,14 @@ if __name__ == '__main__':
     frr.write('Filename,ContaminationFraction,KernelScale,Lambda,NumExpansionDimensions,StandardizeData,BetaTolerance,BetaTolerance,GradientTolerance,IterationLimit\n')
     frr.close()
     
-    frr=open("Results/MatOCSVM_Uni.csv", "w")
-    frr.write('Filename,ContaminationFraction,KernelScale,Lambda,NumExpansionDimensions,StandardizeData,BetaTolerance,BetaTolerance,GradientTolerance,IterationLimit\n')
-    frr.close()
-    
-    frr=open("Results/MatOCSVM_Bi.csv", "w")
-    frr.write('Filename,ContaminationFraction,KernelScale,Lambda,NumExpansionDimensions,StandardizeData,BetaTolerance,BetaTolerance,GradientTolerance,IterationLimit\n')
-    frr.close()
+    if os.path.exists("Results/MatOCSVM_Uni.csv") == 0:
+        frr=open("Results/MatOCSVM_Uni.csv", "w")
+        frr.write('Filename,ContaminationFraction,KernelScale,Lambda,NumExpansionDimensions,StandardizeData,BetaTolerance,BetaTolerance,GradientTolerance,IterationLimit\n')
+        frr.close()
+    if os.path.exists("Results/MatOCSVM_Bi.csv") == 0:
+        frr=open("Results/MatOCSVM_Bi.csv", "w")
+        frr.write('Filename,ContaminationFraction,KernelScale,Lambda,NumExpansionDimensions,StandardizeData,BetaTolerance,BetaTolerance,GradientTolerance,IterationLimit\n')
+        frr.close()
     
     for fname in master_files:
         ocsvm(fname, parameters, 0)
